From Elm to a Simple Web
Code refactoring, changing the strategy. Previously, I would have a
"single-page application" type website, but now I will do everything
following the principles of a lighter and cleaner Internet.

Initially, I had the idea of using Elm for everything from now on, but
since I started following Luke Smith and others, I remembered the number
of problems we currently have on the Internet that are caused by the
overuse of JavaScript; my idea now is to find ways to create the
lightest and most beautiful websites and tools possible, using
JavaScript only when absolutely necessary.