Restructuring the website
I could call this an initial commit if I were using a virgin
repository. Since I like to use tools in unconventional ways, I will
make this mess without remorse.

Now that I need to establish my online presence as quickly as possible,
after all, time has passed and it's been a long time, I am restructuring
my entire approach to website development.

For my personal website, I will adopt a very, very minimalist and
incremental approach, meaning I will start with the bare minimum and
just a little styling, and over time I will evolve the website until it
is as pleasant as possible to use. If I could summarize the real
intention in a single sentence, it would be: I will not put the cart
before the horse.

In this commit, the website is not ready for use, but I believe that the
basic structure is well defined, so from now on I will start working on
filling in the content that I plan to make available on my website.