True first version in production
After a lot of delays, my new personal website is live.

It's really interesting how much time I'm spending doing something that
potentially no one will ever see. However, that's how I deal with
things, so how can I help myself, right?

I would have liked to add an introductory post, but I'll settle for
using *I Am Not a Superman* for that.

I believe that in the next commit I will already bring numerous changes
to the project structure, because the way it is now is not easily
scalable.