refactor: PageHeader styled; global.css; layouts
I'm not going to try to make my website more beautiful. Its
ultra-minimalist look is enough for now, especially since I just want to
migrate to Vercel and be able to disable the VPS on Vultr.

In this first styling commit, I'm starting to understand where I should
put each style rule. I created a `src/styles/global.css`, based on the
Astro documentation, where I'll put all the rules that are truly
global. For example, rules for defining a consistent starting point
across browsers.

Everything that is not global will be styled within the components. This
is a very simple website, so there's no much room to be creative. The
only novelty that this refactoring is bringing is that I'm adopting more
responsive approaches when styling the components, such as using `em`
and `rem` for size units.

Finally, I refactored the layout files to Markdown, where I was
unnecessarily repeating what was already perfectly defined in
`BaseLayout.astro`.