refactor: non-selectable elements; end of refactoring
With this commit, I finally complete the refactoring in Astro.

Here, I adjusted so that only the main content is selectable by the
user, since structural elements are just that, structural.

I also took the opportunity to improve the behavior of pages in the
browsers' reading mode. Here I learned an interesting technique: add the
`hidden` attribute in the HTML for elements that you don't want to be
displayed in reading mode, but styled with `display: block` (or whatever
is more appropriate), so that the browser displays them normally.